First Assignment
We were given a pick-your-project, assignment for the first installment of Object Oriented Java. I decided to do the Currency Conversion challenge. I originally created a class (ConvertMoney) that had a generic method for converting currencies, when given an amount of money, and a conversion rate.

The teacher wanted explicit conversion methods for the assignment. So I rewrote my main code to have these methods (ConvertMultipleCurrenciesStandalone), and created a separate main (ConvertMultipleCurrencies) and class (ConvertingCurrencies) to display the use of static classes. ConvertMoney will no longer work with either version of the main class.

My main classes (MultipleCurrencies and Standalone) have a branching menu option dependant on the user, include input verification, and enable the user to switch menus (convert from and convert to USD) once in a specific menu. The Ok, Cancel and X buttons in JOptionPane work through the input verification.