CourseWork for Object Oriented Java 2
SortSearchAssign, AllSorts and AllSearch written by the professor. We only had to compare the run times between different sorting and searching algorithms, then comment on a discussion post about which was a better alternative, and conversely which was worse. SportsTeam1, and SportsTeam2, also written by professor. We had to create a program that counted the curly braces in each file.